migrate away from server.py for readability
